move factory to infra
