+++ TESTING.md
@@ -1,15 +1,13 @@
-Testing
-====
+# Testing
 
 The purpose of this document is to list features to check when major code change is made.
 
-All-in-one bookmarklet
-----
+## All-in-one bookmarklet
 
 - On JP maimai net, analyze self rating.
 - On international maimai net, analyze friend rating.
 - Play record list: Summary table should display.
-- Play record list: Each play record should contain DX score percentage.
+- Play record list: Each play record should contain DX score percentage and chart constant.
 - Single play record: DX score percentage and chart level should be displayed.
 - Single play record: can analyze score in classic-layout.
 - Self score sorting: by achievement, by ap/fc, by sync, by official level, by internal level, by dx star.
@@ -18,21 +16,18 @@ All-in-one bookmarklet
 - Album page: each photo title becomes clickable and will download the photo with correct filename (date + song name + difficulty)
 - etc.
 
-Classic layout (DX -> FiNALE score converter)
-----
+## Classic layout (DX -> FiNALE score converter)
 
 - Switch between DX and Cafe MiLK
 - Switch between normal display, minus display, and detail display by clicking on achievement percentage.
 
-Rating visualizer
-----
+## Rating visualizer
 
 - Change scale
 - Change min level
 - Change max level
 - Change language
 
-DX Achievement (from FiNALE score)
-----
+## DX Achievement (from FiNALE score)
 
 - Fill example data
